+++ packages/react/src/MediaFile.tsx
@@ -4,23 +4,27 @@ import {ExternalVideo} from './ExternalVideo.js';
 import {ModelViewer} from './ModelViewer.js';
 import type {MediaEdge as MediaEdgeType} from './storefront-api-types.js';
 import type {PartialDeep} from 'type-fest';
+import type {ModelViewerElement} from '@google/model-viewer/lib/model-viewer.js';
 
-interface MediaFileProps {
+type BaseProps = React.HTMLAttributes<
+  HTMLImageElement | HTMLVideoElement | HTMLIFrameElement | ModelViewerElement
+>;
+export interface MediaFileProps extends BaseProps {
   /** An object with fields that correspond to the Storefront API's [Media object](https://shopify.dev/api/storefront/reference/products/media). */
   data: PartialDeep<MediaEdgeType['node'], {recurseIntoArrays: true}>;
   /** The options for the `Image`, `Video`, or `ExternalVideo` components. */
   mediaOptions?: {
     /** Props that will only apply when an `<Image />` is rendered */
-    image: Omit<ShopifyImageProps, 'data'>;
+    image?: Omit<ShopifyImageProps, 'data'>;
     /** Props that will only apply when a `<Video />` is rendered */
-    video: Omit<React.ComponentProps<typeof Video>, 'data'>;
+    video?: Omit<React.ComponentProps<typeof Video>, 'data'>;
     /** Props that will only apply when an `<ExternalVideo />` is rendered */
-    externalVideo: Omit<
+    externalVideo?: Omit<
       React.ComponentProps<typeof ExternalVideo>['options'],
       'data'
     >;
     /** Props that will only apply when a `<ModelViewer />` is rendered */
-    modelViewer: Omit<typeof ModelViewer, 'data'>;
+    modelViewer?: Omit<typeof ModelViewer, 'data'>;
   };
 }
 
@@ -70,6 +74,7 @@ export function MediaFile({
     }
     case 'Model3d': {
       return (
+        // @ts-expect-error There are issues with the inferred HTML attribute types here for ModelViewer (and contentEditable), but I think that's a little bit beyond me at the moment
         <ModelViewer
           {...passthroughProps}
           {...mediaOptions?.modelViewer}
@@ -78,11 +83,11 @@ export function MediaFile({
       );
     }
     default: {
-      const typenameMissingMessage = `<MediaFile /> requires the '__typename' property to exist on the 'data' prop in order to correctly render the correct component for this media. Rendering 'null' by default`;
+      const typenameMissingMessage = `<MediaFile /> requires the '__typename' property to exist on the 'data' prop in order to render the matching sub-component for this type of media.`;
       if (__HYDROGEN_DEV__) {
         throw new Error(typenameMissingMessage);
       } else {
-        console.error(typenameMissingMessage);
+        console.error(`${typenameMissingMessage}  Rendering 'null' by default`);
         return null;
       }
     }